Fix: Convert keyCode from Long to Int for comparison
Fixed type mismatch error - event.key.keyCode returns Long in Compose,
but we need Int for when expression comparison.

Added .toInt() conversion to fix compilation error.

+++ app/src/main/java/com/rutv/ui/mobile/components/ChannelListItem.kt
@@ -79,7 +79,7 @@ fun ChannelListItem(
     // Handle remote key events
     val onRemoteKeyEvent: (androidx.compose.ui.input.key.KeyEvent) -> Boolean = { event ->
         // Log EVERY key event that reaches this lambda
-        val keyName = when(event.key.keyCode) {
+        val keyName = when(event.key.keyCode.toInt()) {
             19 -> "UP"
             20 -> "DOWN"
             21 -> "LEFT"
